nifty50-analysis-tool/nifty50-analysis-tool/backend/requirements.txt
Django==4.2.7
djongo==1.3.6
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.2.0
celery==5.3.4
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.2
textblob==0.17.1